export_raw_gs_plots

function to generate ungated plots from selected gs node

Usage

export_raw_gs_plots(
  gs,
  node_name,
  channel_x,
  channel_y,
  path_output,
  n_cores = 1,
  x_lab = "x",
  y_lab = "y",
  w_val = 7,
  h_val = 7,
  size_points = 1,
  return_data = FALSE,
  ...
)

Arguments

gs

GatingSet

node_name

TODOLIST

channel_x

TODOLIST

channel_y

TODOLIST

path_output

Path to the directory where to export the plots.

n_cores

Number of cores to use. Default to 1.

x_lab

x-axis label.

y_lab

y-axis label.

w_val

width value. Default to 7 inches.

h_val

height value. Default to 7 inches.

size_points

Size points scatter plot.

return_data

If TRUE, return the list of dataframes used to generate the plots. Default to FALSE.

...

Additional arguments passed to the function

Value

TODOLIST

Examples

A <- 2+2