exports_plots

function to generate plots (no hierarchy) from list of labelled dataframes.

Usage

exports_plots(
  list_gated_data,
  path_output,
  n_cores = 1,
  type_plot = "dens",
  show_legend = TRUE,
  x_lab = "x",
  y_lab = "y",
  size_title_x = 23,
  size_title_y = 23,
  aspect_ratio = NULL,
  w_val = 16,
  h_val = 10,
  size_axis_text = 25,
  export_csv = FALSE,
  side_by_side = FALSE,
  ...
)

Arguments

list_gated_data

list of dataframes. Each dataframe has 3 columns: marker 1 values, marker 2 values and label column.

path_output

Path to the directory where to export the plots.

n_cores

Number of cores to use. Default to 1.

type_plot

the user can choose between density (="dens) or label assignment visualization (="ML")

show_legend

If True it shows the legend for the label assignment visualization. Default to True.

x_lab

x-axis label.

y_lab

y-axis label.

size_title_x

Size x axis label.

size_title_y

Size y axis label.

aspect_ratio

Set aspect ratio. Default to NULL> If = 1, y and x axis ticks have same distance.

w_val

width value. Default to 16.

h_val

height value. Default to 10.

size_axis_text

Size of ticks labels.

export_csv

Export plot data as csv files. Default to False.

side_by_side

Arrange dens plot and ML plot side-by-side. Default to False.

...

Additional arguments passed to the function

Value

TODOLIST

Examples

A <- 2+2