get_train_data

function to import training data based on paths to files.

Usage

get_train_data(
  paths_file = NULL,
  df_paths = NULL,
  n_cores = 1,
  prop_down = NULL,
  remove_class = NULL,
  n_points_per_plot = NULL,
  normalize_data = TRUE,
  vec_col = NULL
)

Arguments

paths_file

Vector of paths. Each path points toward a single csv file containin training info (labels and bivariate expression). paths_file can be also directly the list of dataframes containing labels and bivariate expression.

df_paths

Dataframe containing the paths of file to read. The paths to data must be in the first column. The associated paths to classes are in second column.

n_cores

Number of cores. Default to 1.

prop_down

Proportion of events (downsampling). Default to NULL (downsampling using number of points).

remove_class

Vector of classes to ignore. Default to NULL.

n_points_per_plot

Number of points for downsampling.

normalize_data

If True, data is normalized to 0-1 range. Default to True.

vec_col

vector of columns names if the input dataframes have more than 3 columns. The third column name must always refer to the column with the gate label of each event. Default to NULL.

Value

Dataframe.

Examples

A <- 2+2