magicPlot

function to generate the scatter plot with colored density of the events.

Usage

magicPlot(
  df,
  type = "dens",
  polygons_coords_list = NULL,
  show_legend = TRUE,
  size_axis_text = 18,
  size_title_x = 20,
  size_title_y = 20,
  treat_0_as_gate = FALSE,
  x_lab = "x",
  y_lab = "y",
  gates_to_plot = NULL,
  apply_manual_scale = FALSE,
  hull_only = FALSE,
  size_points = 1,
  concavity_val = 20,
  aspect_ratio = NULL,
  x_lim1 = NULL,
  x_lim2 = NULL,
  y_lim1 = NULL,
  y_lim2 = NULL,
  add_labels = FALSE,
  map_label_polygon = NULL,
  size_pol_name = 6,
  show_marginals = FALSE,
  ...
)

Arguments

df

Dataframe of bivariate markers expression (with labels if gates to plot).

type

Type of plot to generated. "dens"=bivariate density plot. "ML"=events assignments plot

polygons_coords_list

list of gates coordinates. Needed if labels not included in df. Default to NULL.

show_legend

Show legend if type="ML". Default to True.

size_axis_text

Size of axis ticks labels. Default to 18.

size_title_x

Size of x axis label title.

size_title_y

Size of y axis label title.

treat_0_as_gate

Treat 0 label as gate. Defaul to False (0 label is background)

x_lab

Label of x axis.

y_lab

Label off y axis.

gates_to_plot

Select labels to plot.

apply_manual_scale

Apply predifined scale of colors. Default to True.

hull_only

TODOLIST

size_points

Size of points in scatter plot.

concavity_val

Concavity value. Default to 5. Higher value, less jagged boundaries

aspect_ratio

Aspect ratio value. If = 1, y and x axis have equal distance between the ticks labels. Default to NULL.

x_lim1

Minimum limit x axis. Default to NULL.

x_lim2

Max limit x axis. Default to NULL.

y_lim1

Minimum limit y axis. Default to NULL.

y_lim2

Max limit y axis. Default to NULL.

add_labels

add polygon labels. Default to FALSE.

map_label_polygon

map of polygon labels (to assign custom labels). Default to NULL.

size_pol_name

size polygon labels. Default to 6.

show_marginals

show 1D density next to axis. Default to False.

...

Additional arguments passed to the function

Value

ggplot.

Examples

A <- 2+2