magicPred_all

function to predict on plain test data (no hierarchy)

Usage

magicPred_all(
  list_test_data,
  magic_model = NULL,
  ref_model_info = NULL,
  magic_model_n_gates = NULL,
  ref_data_train = NULL,
  prop_down = NULL,
  n_points_per_plot = NULL,
  thr_dist = 0.05,
  n_cores = 1,
  normalize_data = TRUE,
  include_zero_val = TRUE,
  n_cores_all = 1,
  verbose = FALSE,
  seed_n = 40,
  ...
)

Arguments

list_test_data

List of unlabeled dataframes. It has only the two columns of marker expression.

magic_model

Global trained model to predict gates. It can be a single model or list of named models (each model trained on selected number of gates).

ref_model_info

Template model to predic gates.

magic_model_n_gates

Global trained model to predict number of gates. If different from NULL, magic_model is expected to be a list of models to predict certain gates (e.g., 5 models for 2,3,4,5 or 6 gates).

ref_data_train

Template data used to generate ref_model_info. Needed to calculate target-template distance.

prop_down

TODOLIST

n_points_per_plot

Number of points to consider for downsampling. Default to 500.

thr_dist

TODOLIST

n_cores

Number of cores to use to process one sample. Default to 1.

normalize_data

If True, data is normalized to 0-1 range. Default to True.

include_zero_val

considering events labeled as 0 as an additional gate when there is only one gate. Default to True.

n_cores_all

Number of cores to use across all samples. Default to 1.

verbose

If True, print all message and disable tryCatch (any error will stop the execution). Default to False.

seed_n

Numeric for set.seed, default is 40

...

Additional arguments passed to the function

Value

List of Dataframes.

Examples

A <- 2+2